palm/plugins/dbt/requirements.txt
palm >=2.5.1, <3.0
pyyaml >= 5.0
sqlparse >= 0.3.1
semver >= 3.0.1
